ble-daemon/requirements.txt
asyncio
bleak
paho-mqtt
python-dotenv
termcolor
